+++ src/utils/dirname.hpp
@@ -1,0 +1,6 @@
+#include <string>
+
+namespace utils
+{
+std::string dirname(const std::string filename);
+}